+++ oms/src/main/java/com/upedge/oms/modules/pack/request/OrderPackageAddRequest.java
@@ -25,7 +25,7 @@ public class OrderPackageAddRequest{
     /**
     * 
     */
-    private Integer packageNo;
+    private Long packageNo;
     /**
     * 
     */
