+++ oms/src/main/java/com/upedge/oms/modules/pack/request/OrderPackageInfoGetRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class OrderPackageInfoGetRequest {
 
-    private Integer packageNo;
+    private Long packageNo;
 
     private String trackingCode;
 }
